Navigation menu view and placeholder directory, account and OCR views added. Placeholder fragments for directory, account and OCR views added, as well as necessarly xml resources for each.
Added in ActivitySwitchController, which controls the navigation bar, and the layout files for the navigation bar. Attached the fragments to be manipulated by ActivitySwitchController. Switched the redirect from the login placeholder to point to ActivitySwitchController.

Added a recycler view to the placeholder DirectoryAccessController.

Migrated DirectoryActivity into DirectoryAccessController with edits to function with the navigation bar and the fragments system rather than the activity system.

Implemented screen switching with retention of screen state to the navigation bar.

Redirected FileListAdaptor's open folder intent to point to ActivitySwitchController instead.